improve stability & fix websocket deadlock

+++ models/submission.go
@@ -7,16 +7,16 @@ import (
 )
 
 type Submission struct {
-	Sid           string        `gorm:"auto_increment" json:"sid"`
-	Tid           string        `json:"tid"`
-	Uid           string        `json:"uid"`
-	Status        string        `json:"status"`
-	Judge         []byte        `json:"-"`
-	Language      string        `json:"language"`
-	FileName      string        `json:"-"`
-	TimeUsed      uint32        `json:"time_used"`
-	SpaceUsed     uint32        `json:"space_used"`
-	CreatedAt     time.Time     `json:"created_at"`
+	Sid       string    `gorm:"auto_increment" json:"sid"`
+	Tid       string    `json:"tid"`
+	Uid       string    `json:"uid"`
+	Status    string    `json:"status"`
+	Judge     []byte    `json:"-"`
+	Language  string    `json:"language"`
+	FileName  string    `json:"-"`
+	TimeUsed  uint32    `json:"time_used"`
+	SpaceUsed float64   `json:"space_used"`
+	CreatedAt time.Time `json:"created_at"`
 }
 
 type SubmissionExtended struct {
@@ -47,9 +47,9 @@ type SubmissionLite struct {
 }
 
 type JudgeResult struct {
-	Status    string `json:"status"`
-	TimeUsed  uint32 `json:"time_used"`
-	SpaceUsed uint32 `json:"space_used"`
+	Status    string  `json:"status"`
+	TimeUsed  uint32  `json:"time_used"`
+	SpaceUsed float64 `json:"space_used"`
 }
 
 func (s *SubmissionExtended) AfterFind(scope *gorm.Scope) (err error) {
